fix(api): move code event to keel-core
Squashed commit of the following:

commit 4e82b64ed891f8266fb320b05b374c25bb9636ab
Author: Emily Burns <emburns@netflix.com>
Date:   Mon Jul 12 11:13:15 2021 -0700

    fix(api): move code event to keel-core

(cherry picked from commit cc4cc28013fc7b67d629f21f96b1eb03c3af2bf1)
