+++ keel-scm/src/main/kotlin/com/netflix/spinnaker/keel/preview/PreviewEnvironmentCodeEventListener.kt
@@ -13,13 +13,13 @@ import com.netflix.spinnaker.keel.api.PreviewEnvironmentSpec
 import com.netflix.spinnaker.keel.api.Resource
 import com.netflix.spinnaker.keel.api.ResourceSpec
 import com.netflix.spinnaker.keel.api.generateId
-import com.netflix.spinnaker.keel.api.scm.CodeEvent
-import com.netflix.spinnaker.keel.api.scm.CommitCreatedEvent
-import com.netflix.spinnaker.keel.api.scm.PrOpenedEvent
-import com.netflix.spinnaker.keel.api.scm.PrDeclinedEvent
-import com.netflix.spinnaker.keel.api.scm.PrDeletedEvent
-import com.netflix.spinnaker.keel.api.scm.PrEvent
-import com.netflix.spinnaker.keel.api.scm.PrMergedEvent
+import com.netflix.spinnaker.keel.scm.CodeEvent
+import com.netflix.spinnaker.keel.scm.CommitCreatedEvent
+import com.netflix.spinnaker.keel.scm.PrOpenedEvent
+import com.netflix.spinnaker.keel.scm.PrDeclinedEvent
+import com.netflix.spinnaker.keel.scm.PrDeletedEvent
+import com.netflix.spinnaker.keel.scm.PrEvent
+import com.netflix.spinnaker.keel.scm.PrMergedEvent
 import com.netflix.spinnaker.keel.api.titus.TitusClusterSpec
 import com.netflix.spinnaker.keel.docker.ReferenceProvider
 import com.netflix.spinnaker.keel.front50.Front50Cache
