+++ keel-scm/src/main/kotlin/com/netflix/spinnaker/keel/scm/utils.kt
@@ -1,8 +1,6 @@
 package com.netflix.spinnaker.keel.scm
 
 import com.netflix.spectator.api.BasicTag
-import com.netflix.spinnaker.keel.api.scm.CodeEvent
-import com.netflix.spinnaker.keel.api.scm.CommitCreatedEvent
 import com.netflix.spinnaker.keel.front50.model.Application
 import com.netflix.spinnaker.keel.notifications.DeliveryConfigImportFailed
 import org.springframework.context.ApplicationEventPublisher
